Removing -Wsign-conversion warnings
There were quite a few places in the code where we were implicitly
converting signed to unsigned integers or viceversa, specially around
OpenMP-related code.

This commit improves that situation by doing quite a few things:

 * Complete revamp of the OpenMP helper functions, making them safer and
easier for users to use correctly.
 * Functions given to shark's omp_* now receive the thread index as an
unsigned integer rather than it being signed.
 * Removed unused method in merger_tree_reader
 * Adding small assertion to size() methods in ranges.h before
explicitly converting from signed to unsigned integer.
 * Changing the way we seed the random engines so they work well with
all components.

+++ src/main.cpp
@@ -186,7 +186,7 @@ Options read_options(const boost::program_options::variables_map &vm, unsigned i
 #ifdef SHARK_OPENMP
 	threads = vm["threads"].as<unsigned int>();
 	if (threads == 0) {
-		threads = omp_get_max_threads();
+		threads = static_cast<unsigned int>(omp_get_max_threads());
 	}
 #else
 	threads = 1;
